fix: Fix tag order from git tag function.
This list of tags comes in ABC sorted according to git, whereas
last_valid_non_rc_version function assumed it would come in tag
revision order. This manifests itself in finding e.g. 9.9.0 as the
latest revision when 10.0.0 is in the list of tags, as well. Because
this tag is used later to calculate the diff, it winds up double
counting commits and therefor doing bumps it should not be.