feat(project_info_task): Add `git_ops.project_info` task.
Adds a project info mix task which will output the app's name and version to stdout in various formats.

This is an extremely simplistic implementation because I didn't want to have to add any dependencies to git_ops.  I find that I am constantly trying to parse this information out of the `mix.exs` file in various CI/CD stages.